## Summary
- 작가 마이페이지 촬영 패키지 편집 화면을 실제 상품 API 기반으로 연결했습니다.
- 상품 등록 폼, 이미지 업로드, 등록 가능 조건, 최대 3개 제한을 ViewModel에서 관리하도록 추가했습니다.
- dev 화면에서 패키지 편집 화면의 0/1/2/3개 상태를 확인할 수 있는 디버그 진입점을 추가했습니다.

## Changes

### navigation
- `MyPagePackageEdit` route에 `photographerId`를 전달하도록 변경했습니다.
- 작가 마이페이지 패키지 편집 버튼에서 촬영 패키지 편집 화면으로 이동하도록 연결했습니다.
- debug build에서만 사용하는 패키지 편집 fixture route를 추가했습니다.

### feature/mypage
- 촬영 패키지 편집 MVI 화면, 상태, Intent, SideEffect, ViewModel을 구현했습니다.
- 기존 상품 목록 조회, 새 상품 등록, 이미지 선택/업로드, 저장 버튼 활성화 조건을 처리했습니다.
- 수정/삭제는 API 미지원 상태라 준비 중 토스트로 차단했습니다.
- 패키지 카드 메타 정보 정렬, typography, kebab 메뉴, bottom sheet 높이를 UI 요구사항에 맞게
조정했습니다.

### core/data, core/domain
- 상품 조회/등록 API, Source, Service, Repository, UseCase를 추가했습니다.
- 상품 DTO를 domain `ShootingPackage`로 매핑하고, 상품 이미지 업로드 use case를 연결했습니다.

### core/ui
- 공통 background text field를 추가했습니다.
- `CommonModalBottomSheet`의 최대 높이 확장 여부를 opt-in 파라미터로 분리했습니다.

### feature/main
- DevScreen에 패키지 편집 화면 fixture 진입 버튼을 추가했습니다.
- 개발 토큰 설정/게스트 전환/로그아웃을 dev 화면 버튼으로 분리했습니다.

## 구현내용


<img width="315" height="683" alt="KakaoTalk_Video_2026-05-18-00-24-44
(1)"
src="https://github.com/user-attachments/assets/052d2a20-f5fb-4b6e-804b-e6d254a45f00"
/>

## Test Scenario

### Dev fixture로 목록 UI 확인
1. 앱 실행 후 Dev 화면으로 진입합니다.
2. `MyPagePackageEdit DEV (패키지 0개)`를 눌러 빈 목록과 `상품을 추가하세요` 카드가 보이는지
확인합니다.
3. `MyPagePackageEdit DEV (패키지 1개)`, `2개`, `3개`를 각각 눌러 등록된 상품 카드 개수가
상태별로 표시되는지 확인합니다.
4. 패키지 3개 상태에서는 `상품을 추가하세요` 카드가 숨겨지는지 확인합니다.

### 상품 등록 폼 확인
1. `MyPagePackageEdit DEV (패키지 0개)` 진입 후 `상품을 추가하세요`를 누릅니다.
2. `새 상품 등록` 폼으로 전환되는지 확인합니다.
3. 상품명을 입력하고 촬영 시간을 선택하기 전에는 저장 버튼이 비활성인지 확인합니다.
4. 촬영 시간을 선택하면 저장 버튼이 활성화되는지 확인합니다.
5. `기타 안내사항` placeholder는 회색이고, 입력한 텍스트는 검정색으로 표시되는지 확인합니다.
6. 뒤로가기를 누르면 목록 화면으로 돌아가는지 확인합니다.

### 기존 상품 카드 액션 확인
1. `MyPagePackageEdit DEV (패키지 1개)` 이상으로 진입합니다.
2. 상품 카드 우측 kebab 메뉴를 누릅니다.
3. 바텀시트가 화면 상단까지 올라가지 않고 `상품 수정하기`, `상품 삭제하기` 두 옵션이 보일 정도만 올라오는지 확인합니다.
4. `상품 수정하기` 또는 `상품 삭제하기`를 누르면 `아직 지원하지 않는 기능입니다.` 토스트가 표시되는지 확인합니다.


## Verification
- `./gradlew ktlintCheck --quiet`
- `./gradlew :app:compileDebugKotlin --quiet`
- `./gradlew :feature:mypage:compileDebugKotlin --quiet`
- `./gradlew :feature:mypage:testDebugUnitTest --quiet`

Closes #183

+++ core/data/src/main/kotlin/com/hm/picplz/data/mapper/PhotographerMapper.kt
@@ -101,7 +101,7 @@ fun ReviewListDto.toReviewData(): PhotographerReviewData {
 
 fun ProductDto.toShootingPackage(): ShootingPackage {
     return ShootingPackage(
-        packageId = productId.toInt(),
+        packageId = productId,
         title = title ?: name.orEmpty(),
         price = price ?: shootPrice ?: 0,
         imageUri = imageUrl ?: productPhotos?.firstOrNull().orEmpty(),
